Identify Text/requirements.txt
torch
transformers
Pillow
accelerate
safetensors
torchvision
numpy
opencv-python
ultralytics
scipy